fix test failing for mysql < 5.6.4
